+++ public/data/feedsData.json
@@ -1,0 +1,18 @@
+[
+  {
+    "id": 1,
+    "userName": "그니"
+  },
+  {
+    "id": 2,
+    "userName": "보니"
+  },
+  {
+    "id": 3,
+    "userName": "효정"
+  },
+  {
+    "id": 4,
+    "userName": "영완"
+  }
+]
